+++ projects/ids-enterprise-ng/src/lib/module-nav-settings/soho-module-nav-settings.component.ts
@@ -92,7 +92,7 @@ export class SohoModuleNavSettingsComponent implements AfterViewInit, AfterViewC
 
   ngAfterViewChecked() {
     if (this.modulenavsettings && this._updateRequired) {
-      this.ngZone.runOutsideAngular(() => this.modulenavsettings?.updated());
+      this.ngZone.runOutsideAngular(() => this.modulenavsettings?.updated(this._options));
       this._updateRequired = false;
     }
   }